mumbling/test_mumbling.py
# PyTest for mumbling.py

from mumbling import accum


def test_0():
    assert(accum('ZpglnRxqenU') ==
           'Z-Pp-Ggg-Llll-Nnnnn-Rrrrrr-Xxxxxxx-Qqqqqqqq-Eeeeeeeee-Nnnnnnnnnn-Uuuuuuuuuuu')


def test_1():
    assert(accum('NyffsGeyylB') ==
           'N-Yy-Fff-Ffff-Sssss-Gggggg-Eeeeeee-Yyyyyyyy-Yyyyyyyyy-Llllllllll-Bbbbbbbbbbb')


def test_2():
    assert(accum('MjtkuBovqrU') ==
           'M-Jj-Ttt-Kkkk-Uuuuu-Bbbbbb-Ooooooo-Vvvvvvvv-Qqqqqqqqq-Rrrrrrrrrr-Uuuuuuuuuuu')


def test_3():
    assert(accum('EvidjUnokmM') ==
           'E-Vv-Iii-Dddd-Jjjjj-Uuuuuu-Nnnnnnn-Oooooooo-Kkkkkkkkk-Mmmmmmmmmm-Mmmmmmmmmmm')


def test_4():
    assert(accum('HbideVbxncC') ==
           'H-Bb-Iii-Dddd-Eeeee-Vvvvvv-Bbbbbbb-Xxxxxxxx-Nnnnnnnnn-Cccccccccc-Ccccccccccc')
